Hook new implementation into old API
This commit replaces the old synchronous implementation with the new
asynchonous one under the hood, while still keeping the old API. This
way we can test most parts of the new implementation without having to
implement and adapt to a new API. Note that the new implementation is
inherently asynchronous under the hood and only synchronized in the
top-level request function.

The new API will be implemented later, including support for completely
asynchronous requests.

+++ module/surface_sam_ssh.h
@@ -205,7 +205,7 @@ struct surface_sam_ssh_rqst {
 	u8 iid;				// instance ID
 	u8 chn;				// channel
 	u8 snc;				// expect response flag (bool: 0/1)
-	u8 cdl;				// command data length (length of payload)
+	u16 cdl;			// command data length (length of payload)
 	u8 *pld;			// pointer to payload of length cdl
 };
